Move mock classes to _support
