+++ tests/_support/Autoloader/MockAutoloader.php
@@ -1,0 +1,24 @@
+<?php namespace CodeIgniter\Autoloader;
+
+class MockAutoloader extends Autoloader
+{
+
+	protected $files = [];
+
+	//--------------------------------------------------------------------
+
+	public function setFiles($files)
+	{
+		$this->files = $files;
+	}
+
+	//--------------------------------------------------------------------
+
+	protected function requireFile($file)
+	{
+		return in_array($file, $this->files) ? $file : false;
+	}
+
+	//--------------------------------------------------------------------
+
+}
